fix(lead-sources): split segment-sql client/server, drop dup delete toast, suppress nuqs lint
- Move customerSegments const + CustomerSegment type to constants/
  to keep drizzle-orm and customers schema out of the client bundle
- DangerZone.onDelete no longer adds a per-call onError; the
  mutation-level handler in useLeadSourceActions already toasts
- Suppress two false-positive react-hooks-extra warnings (nuqs
  setTab is not React setState; controlled-form sync on prop change
  is intentional)
