+++ src/shared/entities/lead-sources/constants/customer-segments.ts
@@ -1,0 +1,3 @@
+export const customerSegments = ['all', 'active', 'signed', 'dead'] as const
+
+export type CustomerSegment = (typeof customerSegments)[number]
