+++ src/shared/entities/lead-sources/lib/segment-sql.ts
@@ -1,13 +1,10 @@
 import type { SQL } from 'drizzle-orm'
+import type { CustomerSegment } from '@/shared/entities/lead-sources/constants/customer-segments'
 
 import { and, eq, inArray, sql } from 'drizzle-orm'
 
 import { customers } from '@/shared/db/schema/customers'
 import { isSignedCustomerSql } from '@/shared/entities/customers/lib/signed-customer-sql'
-
-export const customerSegments = ['all', 'active', 'signed', 'dead'] as const
-
-export type CustomerSegment = (typeof customerSegments)[number]
 
 /**
  * SQL predicate for the 4-state customer segmentation used by the lead-source
