+++ src/trpc/routers/lead-sources.router.ts
@@ -15,7 +15,8 @@ import { leadSourcesTable } from '@/shared/db/schema/lead-sources'
 import { meetings } from '@/shared/db/schema/meetings'
 import { proposals } from '@/shared/db/schema/proposals'
 import { isSignedCustomerSql } from '@/shared/entities/customers/lib/signed-customer-sql'
-import { buildSegmentWhere, customerSegments } from '@/shared/entities/lead-sources/lib/segment-sql'
+import { customerSegments } from '@/shared/entities/lead-sources/constants/customer-segments'
+import { buildSegmentWhere } from '@/shared/entities/lead-sources/lib/segment-sql'
 import { leadSourceFormConfigSchema } from '@/shared/entities/lead-sources/schemas'
 import { computeFinalTcp } from '@/shared/entities/proposals/lib/compute-final-tcp'
